refactor all the things
Pass the events traight through to loggly without any kind of name
mangling.

Remove unecessary concurrency in the design. journald is fast and
following the cursor. I don't believe that the JSON parse time is going
to be an issue.

If logfile is "-", log directly to stdout.

Remove all log statements from the libraries.

Don't set the runtime concurrency, use the default one.

Fix the docker builds, build the current code not the one on github.

Handle the failure cases properly so the process doesn't get blocked.

Vendor external deps.